This commit violates my own guidelines for atomic commits. Removes grunt tooling and dependencies and does a super basic conversion to TypeScript.
The project will target the latest TypeScript, version 1.8.0. I've evaluated grunt-ts and some of the other tooling that wraps the TypeScript compiler, and have decided it's best just to run the tsc command directly, or wrap it in npm scripts. Having a compiler that is outside the task runner's tools just muddles things up a bit so I've pulled out Grunt. Trends in blogging and podcasts seem to suggest I wouldn't be the first to make this decision.

External to this repository, I've created a set of type definitions for the Google Charts API at https://github.com/angular-google-chart/google-chart-typings. This project uses [Typings](https://github.com/typings/typings) to manage TypeScript definition dependencies.

The projects current state with the Grunt task runner removed makes it difficult to test, but I'll do a little work to clean that up later.